Tying to get api working
Unable to send requests to api controllers
