+++ BugBlaze.csproj
@@ -16,4 +16,9 @@
     <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="5.0.1" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Pages\Users\" />
+    <Folder Include="Pages\Teams\" />
+  </ItemGroup>
+
 </Project>
