+++ Data/Models/Bug.cs
@@ -13,5 +13,15 @@ namespace BugBlaze.Data.Models
         public string Description { get; set; }
         public string Component { get; set; }
         public int ProjectId { get; set; }
+        public BugSeverity Severity { get; set; }
+
+        public enum BugSeverity
+        {
+            low = 1,
+            medium = 2,
+            high = 3,
+            critical = 4
+        }
     }
+
 }
