+++ Pages/Bugs/EditBug.razor
@@ -2,13 +2,13 @@
 
 @using BugBlaze.Data.Models
 
-@page "/editproject/{id}"
+@page "/editbug/{id}"
 
 @inject CustomHttpClient Http
 @inject NavigationManager NavigationManager
 @inject AppSettingsService AppSettingsService
 
-<h2>Edit Employee</h2>
+<h2>Edit Bug</h2>
 <hr />
 <form>
     <div class="row">
@@ -29,7 +29,7 @@
     </div>
     <div class="row">
         <div class="form-group">
-            <input type="button" class="btn btn-primary" @onclick="@UpdateEmployee" value="Update" />
+            <input type="button" class="btn btn-primary" @onclick="@UpdateBug" value="Update" />
             <input type="button" class="btn" @onclick="@Cancel" value="Cancel" />
         </div>
     </div>
@@ -49,7 +49,7 @@
         project = await Http.GetJsonAsync<Project>(baseUrl + "/api/projectsapi/details/" + id);
     }
 
-    protected async Task UpdateEmployee()
+    protected async Task UpdateBug()
     {
         await Http.PutJsonAsync(baseUrl + "/api/projectsapi/edit/" + id, project);
         NavigationManager.NavigateTo("listprojects");
