+++ Pages/Projects/CreateProject.razor
@@ -1,7 +1,7 @@
 ﻿
 @using BugBlaze.Data.Models
 
-@page "/addproject"
+@page "/createproject"
 
 @inject CustomHttpClient Http
 @inject NavigationManager NavigationManager
