+++ appsettings.json
@@ -10,7 +10,7 @@
     }
   },
   "MySettings": {
-    "BaseUrl": "http://localhost:5001"
+    "BaseUrl": "http://localhost:44397"
   },
   "AllowedHosts": "*"
 }
